Books
This problem was annoying to get the right. It was a challenge to establish an
invariant for the while loop, as well as defining the algorithm to minimize
redundancy. Overall, I think my final solution is as clean as it could be,
though I don't feel that my approach is particularly generalizable.